Trait analytics3::cmn::NestedType

pub trait NestedType: MarkerTrait { }

Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.