Struct analytics3::CustomMetric

pub struct CustomMetric {
    pub index: Option<i32>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub created: Option<String>,
    pub max_value: Option<String>,
    pub min_value: Option<String>,
    pub updated: Option<String>,
    pub web_property_id: Option<String>,
    pub active: Option<bool>,
    pub scope: Option<String>,
    pub parent_link: Option<CustomMetricParentLink>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub account_id: Option<String>,
}

JSON template for Analytics Custom Metric.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

customMetrics.patch (request|response)

customMetrics.insert (request|response)

customMetrics.get (response)

customMetrics.update (request|response)

Fields

index

Index of the custom metric.

kind

Kind value for a custom metric. Set to "analytics#customMetric". It is a read-only field.

name

Name of the custom metric.

created

Time the custom metric was created.

max_value

Max value of custom metric.

min_value

Min value of custom metric.

updated

Time the custom metric was last modified.

web_property_id

Property ID.

active

Boolean indicating whether the custom metric is active.

scope

Scope of the custom metric: HIT or PRODUCT.

parent_link

Parent link for the custom metric. Points to the property to which the custom metric belongs.

type_

Data type of custom metric.

id

Custom metric ID.

self_link

Link for the custom metric

account_id

Account ID.