Struct analytics3::EntityUserLinks

pub struct EntityUserLinks {
    pub kind: Option<String>,
    pub total_results: Option<i32>,
    pub items: Vec<EntityUserLink>,
    pub previous_link: Option<String>,
    pub items_per_page: Option<i32>,
    pub start_index: Option<i32>,
    pub next_link: Option<String>,
}

An entity user link collection provides a list of Analytics ACL links Each resource in this collection corresponds to a single link.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

accountUserLinks.list (response)

webpropertyUserLinks.list (response)

profileUserLinks.list (response)

Fields

kind

Collection type.

total_results

The total number of results for the query, regardless of the number of results in the response.

items

A list of entity user links.

previous_link

Previous link for this account collection.

items_per_page

The maximum number of entries the response can contain, regardless of the actual number of entries returned. Its value ranges from 1 to 1000 with a value of 1000 by default, or otherwise specified by the max-results query parameter.

start_index

The starting index of the entries, which is 1 by default or otherwise specified by the start-index query parameter.

next_link

Next link for this account collection.