Struct analytics3::GaData

pub struct GaData {
    pub contains_sampled_data: Option<bool>,
    pub previous_link: Option<String>,
    pub data_table: Option<GaDataDataTable>,
    pub query: Option<GaDataQuery>,
    pub totals_for_all_results: HashMap<String, String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub rows: Vec<Vec<String>>,
    pub total_results: Option<i32>,
    pub items_per_page: Option<i32>,
    pub next_link: Option<String>,
    pub sample_size: Option<String>,
    pub profile_info: Option<GaDataProfileInfo>,
    pub column_headers: Vec<GaDataColumnHeaders>,
    pub self_link: Option<String>,
    pub sample_space: Option<String>,
}

Analytics data for a given view (profile).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

ga.get data (response)

Fields

contains_sampled_data

Determines if Analytics data contains samples.

previous_link

Link to previous page for this Analytics data query.

data_table

no description provided

query

Analytics data request query parameters.

totals_for_all_results

Total values for the requested metrics over all the results, not just the results returned in this response. The order of the metric totals is same as the metric order specified in the request.

id

Unique ID for this data response.

kind

Resource type.

rows

Analytics data rows, where each row contains a list of dimension values followed by the metric values. The order of dimensions and metrics is same as specified in the request.

total_results

The total number of rows for the query, regardless of the number of rows in the response.

items_per_page

The maximum number of rows the response can contain, regardless of the actual number of rows returned. Its value ranges from 1 to 10,000 with a value of 1000 by default, or otherwise specified by the max-results query parameter.

next_link

Link to next page for this Analytics data query.

sample_size

The number of samples used to calculate the result.

profile_info

Information for the view (profile), for which the Analytics data was requested.

column_headers

Column headers that list dimension names followed by the metric names. The order of dimensions and metrics is same as specified in the request.

self_link

Link to this page.

sample_space

Total size of the sample space from which the samples were selected.